Fix bug where using IAM roles drops specified regions
In the agent_options method, if use_aws_metadata is enabled, the
access_key and access_secret_key are set to "". Unfortunately, the value
for aws['regions'] is dropped from the config hash. This commit puts it
back in, and sets it to the value specified by the user in the plugin
config yaml file.

Signed-off-by: Mike McClurg <mike.mcclurg@gmail.com>